вынос отложенного таймера из плеера
