+++ project/src/components/list-films/list-films.tsx
@@ -1,4 +1,3 @@
-import { BaseSyntheticEvent, useState } from 'react';
 
 import { Film } from '../../types/types';
 
@@ -11,21 +10,9 @@ type ListFilmProps = {
 function ListFilm(props: ListFilmProps): JSX.Element {
   const { films } = props;
 
-  const [activeId, setActiveId] = useState<string | null>(null);
-
-  const handleMouseOver = (evt: BaseSyntheticEvent) => {
-    const target = evt.target as Element;
-    const parent = target.parentElement as Element;
-    if (parent.classList.contains('small-film-card') || target.tagName === 'A') {
-      setActiveId(parent.id);
-    } else {
-      setActiveId(null);
-    }
-  };
-
   return (
-    <div className="catalog__films-list" onMouseOver={handleMouseOver} onMouseOut={() => setActiveId(null)}>
-      {films.map((film) => <SmallFilmCard key={film.id} film={film} isActive={film.id.toString() === activeId}/>)}
+    <div className="catalog__films-list">
+      {films.map((film) => <SmallFilmCard key={film.id} film={film} />)}
     </div>
   );
 }
